Added mappers to file structure
