+++ public/rss.xml
@@ -4,7 +4,7 @@
         <title>Evgenii Grigorev</title>
         <link>https://eagrigorev.com</link>
         <description>Software developer and artist based in Thessaloniki, Greece.</description>
-        <lastBuildDate>Tue, 01 Oct 2024 20:13:58 GMT</lastBuildDate>
+        <lastBuildDate>Tue, 01 Oct 2024 20:33:13 GMT</lastBuildDate>
         <docs>https://validator.w3.org/feed/docs/rss2.html</docs>
         <generator>https://github.com/jpmonette/feed</generator>
         <language>en</language>
